requirements.txt
PyQt6==6.7.0
keyboard==0.13.5
darkdetect==0.8.0
Pillow==10.4.0
packaging==24.0
